moved work from open_data to shared utils, finding missing d2 routes
