+++ gtfs_digest/_transit_routes_on_shn.py
@@ -185,7 +185,7 @@ def routes_shn_intersection(buffer_amount: int, file_name: str) -> gpd.GeoDataFr
     )
 
     # Clean up
-    gdf2.District = gdf2.District.fillna(0).astype(int)
+    # gdf2.District = gdf2.District.fillna(0).astype(int)
     return gdf2
 
 def group_route_district(df: pd.DataFrame, pct_route_on_hwy_agg: str) -> pd.DataFrame:
@@ -247,7 +247,7 @@ def dissolve_buffered_for_map(buffer_amount: str) -> gpd.GeoDataFrame:
     gdf = gpd.read_parquet(HWY_FILE, storage_options={"token": credentials.token})
 
     # Dissolve by district
-    gdf2 = gdf.dissolve("District").reset_index()[["geometry", "District", "shn_route"]]
+    gdf2 = gdf.dissolve("District").reset_index()[["geometry", "district", "shn_route"]]
 
     # Save
     gdf2.to_parquet(
@@ -273,11 +273,11 @@ def final_transit_route_shs_outputs(
     # Filter out for any pct_route_on_hwy that we deem too low & for the relevant district.
     open_data_df = open_data_df.loc[
         (open_data_df.pct_route_on_hwy_across_districts > pct_route_intersection)
-        & (open_data_df.District.str.contains(district))
+        #& (open_data_df.District.str.contains(district))
     ]
     # intersecting_gdf.District = intersecting_gdf.District
     intersecting_gdf = intersecting_gdf.loc[
-        intersecting_gdf.District.astype(str).str.contains(district)
+        intersecting_gdf.district.astype(str).str.contains(district)
     ]
 
     # Join back to get the original transit route geometries and the names of the
